Refactored utils under the google/gmp/utils directory.
