+++ operators/dv360/dv360_download_report_by_file_path.py
@@ -17,7 +17,7 @@
 ###########################################################################
 
 
-from gcs_hook import GoogleCloudStorageHook
+from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
 from airflow import models
 
 import tempfile
