+++ operators/dv360/dv360_erf_upload_bq.py
@@ -21,7 +21,7 @@ from airflow import models
 from airflow.contrib.hooks.bigquery_hook import BigQueryHook
 from airflow.contrib.hooks.bigquery_hook import BigQueryBaseCursor
 from airflow.contrib.hooks.gcs_hook import GoogleCloudStorageHook
-from utils.download_and_transform_erf import download_and_transform_erf
+from google.gmp.utils.download_and_transform_erf import download_and_transform_erf
