+++ operators/dv360/dv360_run_query_operator.py
@@ -16,7 +16,7 @@
 #
 ###########################################################################
 
-from hooks.dbm_hook import DV360Hook
+from google.gmp.hooks.gmp_dv360_hook import DisplayVideo360Hook
 from airflow import models
 
 
@@ -36,7 +36,7 @@ class DV360RunQueryOperator(models.BaseOperator):
 
   def execute(self, context):
     if(self.service == None):
-      hook = DV360Hook(dbm_conn_id=self.conn_id)
+      hook = DisplayVideo360Hook(gcp_conn_id=self.conn_id)
       self.service = hook.get_service()
 
     request = self.service.queries().runquery(queryId=self.query_id, body={})
